requirements.txt
Flask
Flask-Cors
Gunicorn
openai
python-dotenv
google-serp-api
google-search-results
pytrends